requirements.txt
Flask
python_dotenv
langchain_groq